move large test into separate file and rename mod to testing
